Correct column product table type

+++ database/migrations/2016_10_26_135938_create_product_table.php
@@ -19,7 +19,7 @@ class CreateProductTable extends Migration
             $table->date('dateExpiration');
             $table->tinyInteger('exclude')->default(0);
             $table->integer('productId')->unsigned()->nullable();
-            $table->integer('SphCorrected')->unsigned();
+            $table->double('SphCorrected')->unsigned();
             $table->timestamps();
         });
     }
